Update inboxscanner.gs
Much more optimized version. As opposed to looping through the emails, I am now storing the email sender in an array and then checking each row of data in the spreadsheet to see if each email exists. I was able to compare 8000 rows of email addresses with 300 emails in my inbox in 22 seconds!! The earlier version of the code could only get through 300 emails in 4 minutes!